requirements.txt
tensorflow==2.1.0
cloudml-hypertune==0.1.0.dev6
